+++ src/app/index.tsx
@@ -1,4 +1,4 @@
-import { ConfigProvider } from 'antd';
+import { ConfigProvider, theme } from 'antd';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 
@@ -18,7 +18,14 @@ if (domNode) {
   const root = createRoot(domNode);
   root.render(
     <Provider store={store}>
-      <ConfigProvider componentSize="small">
+      <ConfigProvider
+        componentSize="small"
+        theme={{
+          algorithm: window.matchMedia('(prefers-color-scheme: dark)').matches
+            ? theme.darkAlgorithm
+            : theme.defaultAlgorithm,
+        }}
+      >
         <App />
       </ConfigProvider>
     </Provider>,