+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.0.0-beta.4]
 ### added
 - new feature of invisible borders per app
 - new easy way to hard restart the services and AHK